+++ CCD_prototype/CCD/makeintmat.h
@@ -223,9 +223,9 @@ public:
     unsigned long int               Identity_ppph(int p1, int p2, int p3, int h1);
     unsigned long int               Identity_hhhppp(int h1, int h2, int h3, int p1, int p2, int p3);
     unsigned long int               Identity_hhhhhp(int h1, int h2, int h3, int h4, int h5, int p1); //this is a special function for T3e alone
-    spp::sparse_hash_map<int, double>           Vhhhp_elements; //needed for T3
-    spp::sparse_hash_map<int, double>           Vhhpp_elements;
-    spp::sparse_hash_map<int, double>           Vppph_elements; //needed for T3
+    spp::sparse_hash_map<unsigned long int, double>           Vhhhp_elements; //needed for T3
+    spp::sparse_hash_map<unsigned long int, double>           Vhhpp_elements;
+    spp::sparse_hash_map<unsigned long int, double>           Vppph_elements; //needed for T3
 
 
     //Eigen::VectorXd                 Vhhpp_vector;
